Restructure test folder
