--- app/src/main/java/com/andela/movit/background/InfiniteService.java
@@ -1,75 +1,0 @@
-package com.andela.movit.background;
-
-import android.app.Service;
-import android.content.Intent;
-import android.os.Handler;
-import android.os.HandlerThread;
-import android.os.IBinder;
-import android.os.Looper;
-import android.os.Message;
-import android.support.annotation.WorkerThread;
-
-public abstract class InfiniteService extends Service {
-
-    private volatile Looper mServiceLooper;
-    private volatile ServiceHandler mServiceHandler;
-    private String mName;
-    private boolean mRedelivery;
-
-    private final class ServiceHandler extends Handler {
-        public ServiceHandler(Looper looper) {
-            super(looper);
-        }
-
-        @Override
-        public void handleMessage(Message msg) {
-            onHandleIntent((Intent)msg.obj);
-        }
-    }
-
-    public InfiniteService(String name) {
-        super();
-        mName = name;
-    }
-
-    public void setIntentRedelivery(boolean enabled) {
-        mRedelivery = enabled;
-    }
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-        HandlerThread thread = new HandlerThread("IntentService[" + mName + "]");
-        thread.start();
-
-        mServiceLooper = thread.getLooper();
-        mServiceHandler = new ServiceHandler(mServiceLooper);
-    }
-
-    @Override
-    public void onStart(Intent intent, int startId) {
-        Message msg = mServiceHandler.obtainMessage();
-        msg.arg1 = startId;
-        msg.obj = intent;
-        mServiceHandler.sendMessage(msg);
-    }
-
-    @Override
-    public int onStartCommand(Intent intent, int flags, int startId) {
-        onStart(intent, startId);
-        return mRedelivery ? START_REDELIVER_INTENT : START_NOT_STICKY;
-    }
-
-    @Override
-    public void onDestroy() {
-        mServiceLooper.quit();
-    }
-
-    @Override
-    public IBinder onBind(Intent intent) {
-        return null;
-    }
-
-    @WorkerThread
-    protected abstract void onHandleIntent(Intent intent);
-}
