--- app/src/main/java/com/andela/movit/dialogs/MinuteCallback.java
@@ -1,5 +1,0 @@
-package com.andela.movit.dialogs;
-
-public interface MinuteCallback {
-    void onMinutesSelected(int minutes);
-}
