--- app/src/main/java/com/andela/movit/listeners/IncomingStringCallback.java
@@ -1,5 +1,0 @@
-package com.andela.movit.listeners;
-
-public interface IncomingStringCallback {
-    void onStringArrive(String incomingString);
-}
