--- app/src/main/java/com/andela/movit/listeners/LocationCallback.java
@@ -1,7 +1,0 @@
-package com.andela.movit.listeners;
-
-import com.andela.movit.models.Movement;
-
-public interface LocationCallback {
-    void onLocationDetected(Movement movement);
-}
