+++ app/src/main/java/com/andela/movit/location/IncomingStringCallback.java
@@ -1,0 +1,5 @@
+package com.andela.movit.location;
+
+public interface IncomingStringCallback {
+    void onStringArrive(String incomingString);
+}
