+++ app/src/main/java/com/andela/movit/location/LocationCallback.java
@@ -1,0 +1,7 @@
+package com.andela.movit.location;
+
+import com.andela.movit.models.Movement;
+
+public interface LocationCallback {
+    void onLocationDetected(Movement movement);
+}
