+++ app/src/main/java/com/andela/movit/views/activities/SplashActivity.java
@@ -1,0 +1,16 @@
+package com.andela.movit.views.activities;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+
+import com.andela.movit.R;
+
+public class SplashActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_splash);
+    }
+
+}
