+++ app/src/main/java/com/andela/movit/views/dialogs/MinuteCallback.java
@@ -1,0 +1,5 @@
+package com.andela.movit.views.dialogs;
+
+public interface MinuteCallback {
+    void onMinutesSelected(int minutes);
+}
